+++ runtime/base/bit_vector.cc
@@ -308,7 +308,7 @@ void BitVector::DumpHelper(const char* prefix, std::ostringstream& buffer) const
 
   buffer << '(';
   for (size_t i = 0; i < storage_size_ * kWordBits; i++) {
-    buffer << IsBitSet(i);
+    buffer << IsBitSet(static_cast<uint32_t>(i));
   }
   buffer << ')';
 }
