+++ runtime/base/mutex-inl.h
@@ -40,7 +40,7 @@ namespace art {
 
 #if ART_USE_FUTEXES
 static inline int futex(volatile int *uaddr, int op, int val, const struct timespec *timeout, volatile int *uaddr2, int val3) {
-  return syscall(SYS_futex, uaddr, op, val, timeout, uaddr2, val3);
+  return static_cast<int>(syscall(SYS_futex, uaddr, op, val, timeout, uaddr2, val3));
 }
 #endif  // ART_USE_FUTEXES
