+++ runtime/gc/space/region_space.h
@@ -381,14 +381,14 @@ class RegionSpace FINAL : public ContinuousMemMapAllocSpace {
       return live_bytes_;
     }
 
-    uint GetLivePercent() const {
+    size_t GetLivePercent() const {
       DCHECK(IsInToSpace());
       DCHECK(!IsLargeTail());
       DCHECK_NE(live_bytes_, static_cast<size_t>(-1));
       DCHECK_LE(live_bytes_, BytesAllocated());
       size_t bytes_allocated = RoundUp(BytesAllocated(), kRegionSize);
       DCHECK_GE(bytes_allocated, 0U);
-      uint result = (live_bytes_ * 100U) / bytes_allocated;
+      size_t result = (live_bytes_ * 100U) / bytes_allocated;
       DCHECK_LE(result, 100U);
       return result;
     }
