+++ runtime/mirror/iftable-inl.h
@@ -25,7 +25,7 @@ namespace mirror {
 inline void IfTable::SetInterface(int32_t i, Class* interface) {
   DCHECK(interface != NULL);
   DCHECK(interface->IsInterface());
-  const size_t idx = i * kMax + kInterface;
+  const int32_t idx = i * kMax + kInterface;
   DCHECK_EQ(Get(idx), static_cast<Object*>(nullptr));
   SetWithoutChecks<false>(idx, interface);
 }
